+++ src/test/java/nextstep/subway/acceptence/SubwayApplicationTests.java
@@ -1,0 +1,12 @@
+package nextstep.subway.acceptence;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SubwayApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+}
