+++ src/test/java/nextstep/subway/acceptence/line/LineAcceptanceFixture.java
@@ -1,4 +1,4 @@
-package nextstep.acceptence.line;
+package nextstep.subway.acceptence.line;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
