+++ src/test/java/nextstep/subway/acceptence/lineStation/LineStationAcceptanceTest.java
@@ -1,9 +1,9 @@
-package nextstep.acceptence.lineStation;
+package nextstep.subway.acceptence.lineStation;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.acceptence.exception.AcceptanceTest;
-import nextstep.acceptence.line.LineAcceptanceFixture;
+import nextstep.subway.acceptence.util.AcceptanceTest;
+import nextstep.subway.acceptence.line.LineAcceptanceFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,8 +11,7 @@ import org.springframework.http.HttpStatus;
 
 import java.sql.SQLException;
 
-import static nextstep.acceptence.lineStation.LineStationAcceptanceFixture.*;
-import static nextstep.acceptence.station.StationAcceptanceFixture.지하철역_생성후_ID_를_리턴한다;
+import static nextstep.subway.acceptence.station.StationAcceptanceFixture.지하철역_생성후_ID_를_리턴한다;
 
 @DisplayName("지하철구간 관련 기능")
 public class LineStationAcceptanceTest extends AcceptanceTest {
@@ -45,11 +44,11 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
     @DisplayName("새로운역을 상행 종점에 등록한다")
     void addUpLineStation() {
         // when
-        ExtractableResponse<Response> response = 지하철_구간을_추가한다(lineId, newStationId, downStationId, 5);
+        ExtractableResponse<Response> response = LineStationAcceptanceFixture.지하철_구간을_추가한다(lineId, newStationId, downStationId, 5);
 
         // then
-        상태코드를_체크한다(response.statusCode(), HttpStatus.CREATED.value());
-        지하철_구간이_추가되었는지_체크한다(response, "선릉역", "역삼역");
+        LineStationAcceptanceFixture.상태코드를_체크한다(response.statusCode(), HttpStatus.CREATED.value());
+        LineStationAcceptanceFixture.지하철_구간이_추가되었는지_체크한다(response, "선릉역", "역삼역");
     }
 
     /**
@@ -61,11 +60,11 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
     @DisplayName("새로운역을 하행 종점에 등록한다")
     void addDownLineStation() {
         // when
-        ExtractableResponse<Response> response = 지하철_구간을_추가한다(lineId, upStationId, newStationId, 5);
+        ExtractableResponse<Response> response = LineStationAcceptanceFixture.지하철_구간을_추가한다(lineId, upStationId, newStationId, 5);
 
         // then
-        상태코드를_체크한다(response.statusCode(), HttpStatus.CREATED.value());
-        지하철_구간이_추가되었는지_체크한다(response, "강남역", "선릉역");
+        LineStationAcceptanceFixture.상태코드를_체크한다(response.statusCode(), HttpStatus.CREATED.value());
+        LineStationAcceptanceFixture.지하철_구간이_추가되었는지_체크한다(response, "강남역", "선릉역");
     }
 
     /**
@@ -77,10 +76,10 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
     @DisplayName("기존 역 사이 길이보다 크거나 같으면 등록을 할 수 없음")
     void sizeMinAndSameNotAddLineStationA() {
         // when
-        ExtractableResponse<Response> response = 지하철_구간을_추가한다(lineId, upStationId, newStationId, 10);
+        ExtractableResponse<Response> response = LineStationAcceptanceFixture.지하철_구간을_추가한다(lineId, upStationId, newStationId, 10);
 
         // then
-        상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
+        LineStationAcceptanceFixture.상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
     }
 
     /**
@@ -96,10 +95,10 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
         long newStationB = 지하철역_생성후_ID_를_리턴한다("화서역");
 
         // when
-        ExtractableResponse<Response> response = 지하철_구간을_추가한다(lineId, newStationA, newStationB, 5);
+        ExtractableResponse<Response> response = LineStationAcceptanceFixture.지하철_구간을_추가한다(lineId, newStationA, newStationB, 5);
 
         // then
-        상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
+        LineStationAcceptanceFixture.상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
     }
 
     /**
@@ -111,9 +110,9 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
     @DisplayName("상행역과 하행역 둘 다 모두 포함되어있으면 추가 할 수 없다.")
     void bothContainStationError() {
         // when
-        ExtractableResponse<Response> response = 지하철_구간을_추가한다(lineId, upStationId, downStationId, 5);
+        ExtractableResponse<Response> response = LineStationAcceptanceFixture.지하철_구간을_추가한다(lineId, upStationId, downStationId, 5);
 
         // then
-        상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
+        LineStationAcceptanceFixture.상태코드를_체크한다(response.statusCode(), HttpStatus.BAD_REQUEST.value());
     }
 }
