+++ src/test/java/nextstep/subway/acceptence/station/StationAcceptanceFixture.java
@@ -1,4 +1,4 @@
-package nextstep.acceptence.station;
+package nextstep.subway.acceptence.station;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
