+++ src/test/java/nextstep/subway/acceptence/station/StationAcceptanceTest.java
@@ -1,15 +1,15 @@
-package nextstep.acceptence.station;
+package nextstep.subway.acceptence.station;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.acceptence.exception.AcceptanceTest;
+import nextstep.subway.acceptence.util.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-import static nextstep.acceptence.station.StationAcceptanceFixture.*;
+import static nextstep.subway.acceptence.station.StationAcceptanceFixture.*;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
